rework if compilation to not speculate labels
